move trades into actions
