+++ lib/agent0/agent0/hyperdrive/agent/__init__.py
@@ -1,8 +1,8 @@
 """Account and wallet with Hyperdrive specific parts"""
 
-from .hyperdrive_actions import HyperdriveActionType, HyperdriveMarketAction
-from .hyperdrive_agent import HyperdriveAgent
-from .hyperdrive_trades import (
+from .hyperdrive_actions import (
+    HyperdriveActionType,
+    HyperdriveMarketAction,
     add_liquidity_trade,
     close_long_trade,
     close_short_trade,
@@ -11,5 +11,6 @@ from .hyperdrive_trades import (
     redeem_withdraw_shares_trade,
     remove_liquidity_trade,
 )
+from .hyperdrive_agent import HyperdriveAgent
 from .hyperdrive_wallet import HyperdriveWallet, HyperdriveWalletDeltas, Long, Short
 from .trade_result import TradeResult, TradeStatus
