+++ lib/agent0/agent0/hyperdrive/agent/hyperdrive_agent.py
@@ -11,8 +11,13 @@ from fixedpointmath import FixedPoint
 from agent0.base import EthAgent, MarketType
 from agent0.base.policies import BasePolicy
 
-from .hyperdrive_actions import HyperdriveMarketAction
-from .hyperdrive_trades import close_long_trade, close_short_trade, redeem_withdraw_shares_trade, remove_liquidity_trade
+from .hyperdrive_actions import (
+    HyperdriveMarketAction,
+    close_long_trade,
+    close_short_trade,
+    redeem_withdraw_shares_trade,
+    remove_liquidity_trade,
+)
 from .hyperdrive_wallet import HyperdriveWallet
 
 if TYPE_CHECKING:
